+++ back/src/bsdasris/resolvers/mutations/__tests__/publishBsdasri.integration.ts
@@ -8,6 +8,7 @@ import {
   readyToPublishData
 } from "../../../__tests__/factories";
 import type { Mutation } from "@td/codegen-back";
+
 const PUBLISH_DASRI = `
 mutation PublishDasri($id: ID!){
   publishBsdasri(id: $id)  {
@@ -45,6 +46,35 @@ describe("Mutation.publishBsdasri", () => {
         })
       })
     ]);
+  });
+
+  it("should publish a draft dasri without emitterWastePackagings", async () => {
+    // Test new rule: `emitterWastePackagings` is not required for publication anymore
+    const { user, company } = await userWithCompanyFactory("MEMBER");
+    const { company: destination } = await userWithCompanyFactory("MEMBER");
+
+    const dasri = await bsdasriFactory({
+      userId: user.id,
+      opt: {
+        isDraft: true,
+        ...initialData(company),
+        ...readyToPublishData(destination),
+        emitterWastePackagings: []
+      }
+    });
+
+    const { mutate } = makeClient(user); // emitter
+
+    const { data } = await mutate<Pick<Mutation, "publishBsdasri">>(
+      PUBLISH_DASRI,
+      {
+        variables: {
+          id: dasri.id
+        }
+      }
+    );
+    expect(data.publishBsdasri.status).toBe("INITIAL");
+    expect(data.publishBsdasri.isDraft).toBe(false);
   });
 
   it("should publish a draft dasri", async () => {